+++ lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:nourish_app/widgets/app_text_field.dart';
 
 import 'package:provider/provider.dart';
 
@@ -6,10 +7,10 @@ import 'router.dart';
 import 'providers.dart';
 
 void main() {
-  runApp(const MyApp());
+  runApp( MaterialApp(home:ProfileInfoScreen())/*MyApp()*/);
 }
 
-class MyApp extends StatelessWidget {
+/*class MyApp extends StatelessWidget {
   const MyApp({super.key});
 
   @override
@@ -25,4 +26,4 @@ class MyApp extends StatelessWidget {
       ),
     );
   }
-}
+}*/
