+++ lib/widgets/app_text_field.dart
@@ -11,6 +11,9 @@ class AppTextField extends StatefulWidget {
   final bool enabled;
   final Widget? prefixIcon;
   final Widget? suffixIcon;
+  final TextStyle? labelStyle;
+  final TextStyle? hintStyle;
+
   final String? Function(String?)? validator;
 
   const AppTextField({
@@ -23,6 +26,9 @@ class AppTextField extends StatefulWidget {
     this.enabled = true,
     this.prefixIcon,
     this.suffixIcon,
+    this.labelStyle,
+    this.hintStyle,
+
     this.validator,
   }) : super(key: key);
 
@@ -41,59 +47,43 @@ class _AppTextFieldState extends State<AppTextField> {
 
   @override
   Widget build(BuildContext context) {
-    return Container(
-      margin: const EdgeInsets.only(bottom: 12.0),
-      padding: const EdgeInsets.all(16.0),
-      decoration: BoxDecoration(
-        borderRadius: BorderRadius.circular(8.0),
-        color: widget.enabled ? Colors.white : Colors.grey[100],
-        border: Border.all(color: Colors.grey[300]!, width: 1.0),
+    return TextFormField(
+      controller: widget.controller,
+      enabled: widget.enabled,
+      obscureText: _isObscured,
+
+      keyboardType: widget.keyboardType,
+      validator: widget.validator,
+      style: const TextStyle(fontSize: 16, color: Colors.black87),
+      decoration: InputDecoration(
+        labelText: widget.label,
+        hintText: widget.hintText,
+        labelStyle:
+            widget.labelStyle ??
+            const TextStyle(color: Colors.grey, fontSize: 14),
+        hintStyle:
+            widget.hintStyle ??
+            const TextStyle(color: Colors.grey, fontSize: 16),
+        prefixIcon: widget.prefixIcon,
+        suffixIcon: _buildSuffixIcon(),
+        contentPadding: const EdgeInsets.symmetric(
+          horizontal: 16,
+          vertical: 12,
+        ),
+        enabledBorder: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(8),
+          borderSide: const BorderSide(color: Colors.grey),
+        ),
+        disabledBorder: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(8),
+          borderSide: const BorderSide(color: Colors.grey),
+        ),
+        focusedBorder: OutlineInputBorder(
+          borderRadius: BorderRadius.circular(8),
+          borderSide: const BorderSide(color: Colors.blue),
+        ),
+        isDense: true,
       ),
-      child:
-          widget.enabled
-              ? TextFormField(
-                controller: widget.controller,
-                enabled: widget.enabled,
-                obscureText: _isObscured,
-                keyboardType: widget.keyboardType,
-                validator: widget.validator,
-                style: const TextStyle(fontSize: 16, color: Colors.black87),
-                decoration: InputDecoration(
-                  labelText: widget.label,
-                  hintText: widget.hintText,
-                  labelStyle: const TextStyle(color: Colors.grey, fontSize: 14),
-                  hintStyle: const TextStyle(color: Colors.grey, fontSize: 16),
-                  prefixIcon: widget.prefixIcon,
-                  suffixIcon: _buildSuffixIcon(),
-                  border: InputBorder.none,
-                  contentPadding: EdgeInsets.zero,
-                  isDense: true,
-                ),
-              )
-              : Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  if (widget.label != null)
-                    Text(
-                      widget.label!,
-                      style: const TextStyle(
-                        fontSize: 14,
-                        color: Colors.grey,
-                        fontWeight: FontWeight.w400,
-                      ),
-                    ),
-                  if (widget.controller?.text.isNotEmpty == true) ...[
-                    const SizedBox(height: 4),
-                    Text(
-                      widget.controller!.text,
-                      style: const TextStyle(
-                        fontSize: 16,
-                        color: Colors.black87,
-                      ),
-                    ),
-                  ],
-                ],
-              ),
     );
   }
 
@@ -116,177 +106,5 @@ class _AppTextFieldState extends State<AppTextField> {
       );
     }
     return widget.suffixIcon;
-  }
-}
-
-// --- Example usage ---
-
-class ProfileInfoScreen extends StatefulWidget {
-  @override
-  _ProfileInfoScreenState createState() => _ProfileInfoScreenState();
-}
-
-class _ProfileInfoScreenState extends State<ProfileInfoScreen> {
-  final GlobalKey<FormState> _formKey = GlobalKey<FormState>();
-  bool _isEditing = false;
-
-  final TextEditingController _nameController = TextEditingController();
-  final TextEditingController _genderController = TextEditingController();
-  final TextEditingController _dobController = TextEditingController();
-  final TextEditingController _heightController = TextEditingController();
-  final TextEditingController _weightController = TextEditingController();
-  final TextEditingController _activityController = TextEditingController();
-
-  String? _notEmptyValidator(String? value) {
-    if (value == null || value.trim().isEmpty) {
-      return 'This field is required';
-    }
-    return null;
-  }
-
-  void _saveProfile() {
-    if (_formKey.currentState!.validate()) {
-      // Save logic here
-      setState(() {
-        _isEditing = false;
-      });
-      ScaffoldMessenger.of(context).showSnackBar(
-        const SnackBar(content: Text('Profile saved successfully')),
-      );
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Colors.white,
-      body: SafeArea(
-        child: Padding(
-          padding: const EdgeInsets.all(20.0),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              const SizedBox(height: 20),
-
-              // Header with edit icon
-              Row(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                children: [
-                  const Text(
-                    'Your info',
-                    style: TextStyle(
-                      fontSize: 18,
-                      fontWeight: FontWeight.w500,
-                      color: Colors.black87,
-                    ),
-                  ),
-                  GestureDetector(
-                    onTap: () {
-                      setState(() {
-                        _isEditing = !_isEditing;
-                      });
-                    },
-                    child: const Icon(Icons.edit, size: 20),
-                  ),
-                ],
-              ),
-              const SizedBox(height: 20),
-
-              // Form Starts Here
-              Expanded(
-                child: Scrollbar(
-                  thumbVisibility: true,
-                  // Always show scrollbar (optional)
-                  thickness: 4,
-                  radius: const Radius.circular(10),
-                  trackVisibility: true,
-                  interactive: true,
-                  child: SingleChildScrollView(
-                    physics: const BouncingScrollPhysics(),
-                    child: Form(
-                      key: _formKey,
-                      child: Column(
-                        children: [
-                          AppTextField(
-                            controller: _nameController,
-                            label: "Name:",
-                            hintText: "your name",
-                            keyboardType: TextInputType.text,
-                            enabled: _isEditing,
-                            validator: _notEmptyValidator,
-                          ),
-                          AppTextField(
-                            controller: _genderController,
-                            label: "Gender:",
-                            hintText: "your gender",
-                            keyboardType: TextInputType.text,
-                            enabled: _isEditing,
-                            validator: _notEmptyValidator,
-                          ),
-                          AppTextField(
-                            controller: _dobController,
-                            label: "DOB:",
-                            hintText: "DD/MM/YYYY",
-                            keyboardType: TextInputType.datetime,
-
-                            enabled: _isEditing,
-                            validator: _notEmptyValidator,
-                          ),
-                          AppTextField(
-                            controller: _heightController,
-                            label: "Height:",
-                            hintText: "Enter height in cm",
-                            keyboardType: TextInputType.number,
-
-                            enabled: _isEditing,
-                            validator: _notEmptyValidator,
-                          ),
-                          AppTextField(
-                            controller: _weightController,
-                            label: "Weight:",
-                            hintText: "Enter weight in kg",
-                            keyboardType: TextInputType.number,
-
-                            enabled: _isEditing,
-                            validator: _notEmptyValidator,
-                          ),
-                          AppTextField(
-                            controller: _activityController,
-                            label: "Activity Level:",
-                            hintText: "e.g., Sedentary, Active, Very Active",
-                            keyboardType: TextInputType.text,
-
-                            enabled: _isEditing,
-                            validator: _notEmptyValidator,
-                          ),
-                          const SizedBox(height: 20),
-
-                          if (_isEditing)
-                            ElevatedButton(
-                              onPressed: _saveProfile,
-                              child: const Text("Save"),
-                            ),
-                        ],
-                      ),
-                    ),
-                  ),
-                ),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-
-  @override
-  void dispose() {
-    _nameController.dispose();
-    _genderController.dispose();
-    _dobController.dispose();
-    _heightController.dispose();
-    _weightController.dispose();
-    _activityController.dispose();
-    super.dispose();
   }
 }
